+++ api/serializers.py
@@ -1,5 +1,6 @@
 from rest_framework import serializers
 from .models import Food
+from .models import Meal
 
 
 class FoodSerializer(serializers.HyperlinkedModelSerializer):
@@ -9,3 +10,11 @@ class FoodSerializer(serializers.HyperlinkedModelSerializer):
         """Meta class to map serializer's fields with the model fields."""
         model = Food
         fields = ('id', 'name', 'calories')
+
+class MealSerializer(serializers.HyperlinkedModelSerializer):
+    """Serializer to map the Model instance into JSON format."""
+
+    class Meta:
+        """Meta class to map serializer's fields with the model fields."""
+        model = Meal
+        fields = ('id', 'name')
