+++ api/tests/test_foods.py
@@ -1,10 +1,9 @@
 from django.test import TestCase
-from .models import Food
+from api.models import Food
+# from .models import Meal
 from rest_framework.test import APIClient
 from rest_framework import status
-from django.urls import reverse
-from rest_framework.test import APIRequestFactory
-from .views import FoodViews
+from api.views import FoodViews
 
 class ModelTestCase(TestCase):
     """This class defines the test suite for the food model."""
@@ -79,25 +78,3 @@ class FoodViewsTest(TestCase):
         food = Food.objects.last()
         response = self.client.delete(f'/api/v1/foods/{food.id}')
         self.assertEquals(response.status_code, status.HTTP_204_NO_CONTENT)
-
-
-class MealViewsTest(TestCase):
-    """Test suite for the meal api views."""
-
-    def setUp(self):
-        """Define the test client and other test variables."""
-        self.client = APIClient()
-        self.breakfast = Meal.objects.create(name="breakfast")
-        self.snack = Meal.objects.create(name="snack")
-        self.lunch = Meal.objects.create(name="lunch")
-        self.dinner = Meal.objects.create(name="dinner")
-
-    def test_gets_all_foods(self):
-        response = self.client.get('/api/v1/meals/')
-        js = response.json()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(js), 4)
-        self.assertEqual(js[0]["name"], self.breakfast.name)
-        self.assertEqual(js[1]["name"], self.snack.name)
-        self.assertEqual(js[2]["name"], self.lunch.name)
-        self.assertEqual(js[3]["name"], self.dinner.name)
